requirements.txt
requests
beautifulsoup4
readability-lxml
lxml
pandas